[lang] Major clean ups in preparation of greater refactorings
