+++ lang/src/lib.rs
@@ -18,20 +18,8 @@
 
 extern crate proc_macro;
 
-use proc_macro::TokenStream;
-use proc_macro2::TokenStream as TokenStream2;
-use quote::ToTokens;
-
-#[proc_macro]
-pub fn contract(input: TokenStream) -> TokenStream {
-    match contract_gen_impl(input) {
-        Ok(tokens) => tokens,
-        Err(err) => err.into_token_stream().into(),
-    }
-}
-
 #[macro_use]
-mod errors;
+mod error;
 
 #[cfg(feature = "generate-api-description")]
 mod api;
@@ -45,35 +33,14 @@ mod parser;
 #[cfg(test)]
 mod tests;
 
-use errors::Result;
+mod contract;
 
-/// Simple wrapper from `proc_macro` to `proc_macro2` and back again.
-///
-/// # Note
-///
-/// The actual `proc_macro` interface has to operate on `proc_macro::TokenStream`
-/// but to keep this library testable we want to use only `proc_macro2::*` entities
-/// internally.
-fn contract_gen_impl(input: TokenStream) -> Result<TokenStream> {
-    contract_gen_impl2(input.into()).map(Into::into)
+use proc_macro::TokenStream;
+
+#[proc_macro]
+pub fn contract(input: TokenStream) -> TokenStream {
+    contract::generate(input.into()).into()
 }
 
-/// Parses the given token stream as pDSL contract, performs some checks and returns
-/// the corresponding contract as token stream.
-pub(crate) fn contract_gen_impl2(input: TokenStream2) -> Result<TokenStream2> {
-    let ast_contract = parser::parse_contract(input.clone())?;
-    let hir_contract = hir::Contract::from_ast(&ast_contract)?;
-    generate_api_description(&hir_contract)?;
-    let tokens = gen::generate_code(&hir_contract);
-    Ok(tokens)
-}
-
-#[cfg(feature = "generate-api-description")]
-fn generate_api_description(contract: &hir::Contract) -> Result<()> {
-    api::generate_api_description(&contract)
-}
-
-#[cfg(not(feature = "generate-api-description"))]
-fn generate_api_description(_contract: &hir::Contract) -> Result<()> {
-    Ok(())
-}
+#[cfg(test)]
+pub use contract::generate_or_err;
